+++ src/test/java/roomescape/service/ReservationServiceTest.java
@@ -1,6 +1,5 @@
 package roomescape.service;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.time.LocalDate;
@@ -14,7 +13,6 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.test.context.jdbc.Sql;
 import roomescape.dto.reservation.ReservationRequest;
-import roomescape.dto.reservation.ReservationResponse;
 import roomescape.dto.reservationtime.ReservationTimeResponse;
 import roomescape.dto.theme.ThemeResponse;
 
@@ -89,46 +87,5 @@ class ReservationServiceTest {
         //when, then
         assertThatThrownBy(() -> reservationService.addReservation(reservationRequest))
                 .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    void 존재하지_않는_id로_삭제할_경우_예외_발생() {
-        //given
-        List<ReservationResponse> allReservations = reservationQueryService.getAllReservations();
-        Long notExistIdToFind = allReservations.size() + 1L;
-
-        //when, then
-        assertThatThrownBy(() -> reservationService.deleteReservation(notExistIdToFind))
-                .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    void 예약_삭제_시_예약_대기가_존재한다면_우선순위가_제일_높은_예약_대기를_자동_승인() {
-        // given
-        List<ReservationResponse> beforeReservations = reservationQueryService.getAllReservations();
-        ReservationResponse beforeFirstReservation = beforeReservations.get(0);
-
-        // when
-        reservationService.deleteReservation(beforeFirstReservation.id());
-
-        // then
-        List<ReservationResponse> afterReservations = reservationQueryService.getAllReservations();
-        ReservationResponse afterFirstReservation = afterReservations.get(0);
-        assertThat(beforeFirstReservation).isNotEqualTo(afterFirstReservation);
-    }
-
-    @Test
-    void 예약_삭제_시_예약_대기가_존재하지_않으면_예약_정상_삭제() {
-        // given
-        List<ReservationResponse> reservations = reservationQueryService.getAllReservations();
-        ReservationResponse firstReservations = reservations.get(0);
-        reservationService.deleteReservation(firstReservations.id());
-
-        // when
-        reservationService.deleteReservation(firstReservations.id());
-
-        // then
-        List<ReservationResponse> afterReservations = reservationQueryService.getAllReservations();
-        assertThat(afterReservations).isEmpty();
     }
 }
